backend/requirements.txt
Django>=5.0,<6.0
django-ninja>=1.1.0,<2.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
psycopg[binary]>=3.2,<4.0
python-dotenv>=1.0,<2.0
whitenoise>=6.7,<7.0
gunicorn>=22.0,<23.0
pytest>=8.0,<9.0
httpx>=0.27,<0.29
azure-cognitiveservices-speech>=1.37,<2.0
pydub>=0.25.0

